fix(upload, user): detect if file existe and get a default image for avatar or banner if no exist and delete file on upload work

+++ backend/user-API/src/preferences/preferences.controller.ts
@@ -31,8 +31,9 @@ export async function updateAvatarPreferences(req: FastifyRequest<{
 	if (!fetchUrl.ok) {
 		return rep.code(fetchUrl.status).send(info);
 	}
-	if (oldPreferences.avatar && oldPreferences.avatar !== 'default.png') {
-		const data = await fetch('http://' + process.env.UPLOAD_IP + '/internal/avatar/' + oldPreferences.avatar, {
+	if (oldPreferences.avatar && oldPreferences.avatar !== 'default.png' && !oldPreferences.avatar.includes('googleusercontent.com')) {
+		const oldAvatar = oldPreferences.avatar.substring(oldPreferences.avatar.lastIndexOf('/') + 1);
+		const data = await fetch('http://' + process.env.UPLOAD_IP + '/internal/avatar/' + oldAvatar, {
 			method: 'DELETE'
 		});
 		console.log(await data.json());
@@ -69,7 +70,8 @@ export async function updateBannerPreferences(req: FastifyRequest<{
 		return rep.code(fetchUrl.status).send(info);
 	}
 	if (oldPreferences.banner && oldPreferences.banner !== 'default.png') {
-		await fetch('http://' + process.env.UPLOAD_IP + '/internal/banner/' + oldPreferences.banner, {
+		const oldBanner = oldPreferences.banner.substring(oldPreferences.banner.lastIndexOf('/') + 1);
+		await fetch('http://' + process.env.UPLOAD_IP + '/internal/banner/' + oldBanner, {
 			method: 'DELETE'
 		});
 	}
